Move Filter View Traits (#2189)
* Initial Migration for FilterView Traits

* Fix styling
