+++ src/Views/Filters/LivewireComponentFilter.php
@@ -3,8 +3,7 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Filters;
 
 use Rappasoft\LaravelLivewireTables\Views\Filter;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasWireables;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\IsLivewireComponentFilter;
+use Rappasoft\LaravelLivewireTables\Views\Filters\Traits\{HasWireables, IsLivewireComponentFilter};
 
 class LivewireComponentFilter extends Filter
 {
