+++ src/Views/Filters/MultiSelectDropdownFilter.php
@@ -4,8 +4,7 @@ namespace Rappasoft\LaravelLivewireTables\Views\Filters;
 
 use Illuminate\Support\Collection;
 use Rappasoft\LaravelLivewireTables\Views\Filter;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasWireables;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasOptions, IsArrayFilter};
+use Rappasoft\LaravelLivewireTables\Views\Filters\Traits\{HasOptions, HasWireables, IsArrayFilter};
 
 class MultiSelectDropdownFilter extends Filter
 {
