+++ src/Views/Filters/NumberFilter.php
@@ -3,8 +3,7 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Filters;
 
 use Rappasoft\LaravelLivewireTables\Views\Filter;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasWireables;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{IsNumericFilter};
+use Rappasoft\LaravelLivewireTables\Views\Filters\Traits\{HasWireables, IsNumericFilter};
 
 class NumberFilter extends Filter
 {
