+++ src/Views/Filters/NumberRangeFilter.php
@@ -3,8 +3,7 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Filters;
 
 use Rappasoft\LaravelLivewireTables\Views\Filter;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasWireables;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasOptions};
+use Rappasoft\LaravelLivewireTables\Views\Filters\Traits\{HasOptions, HasWireables};
 
 class NumberRangeFilter extends Filter
 {
