+++ src/Views/Filters/Traits/HasWireables.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Views\Filters\Traits;
+
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasWireables as HasWireablesCore;
+
+trait HasWireables
+{
+    use HasWireablesCore;
+}
