+++ src/Views/Filters/Traits/IsFilter.php
@@ -1,13 +1,10 @@
 <?php
 
-namespace Rappasoft\LaravelLivewireTables\Views\Traits;
+namespace Rappasoft\LaravelLivewireTables\Views\Filters\Traits;
 
 use Rappasoft\LaravelLivewireTables\Traits\Core\HasLocalisations;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\FilterConfiguration;
+use Rappasoft\LaravelLivewireTables\Views\Filters\Traits\Styling\{HandlesFilterInputAttributes, HandlesFilterLabelAttributes};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasConfig, HasLabelAttributes, HasView};
-use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasCustomPosition, HasVisibility};
-use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\FilterHelpers;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Styling\{HandlesFilterInputAttributes, HandlesFilterLabelAttributes};
 
 trait IsFilter
 {
